requirements.txt
nltk
yfinance==0.2.28
python-dotenv
flask==2.3.3
tweepy
pandas
numpy
werkzeug